requirements_cpu.txt
numpy==1.18.5
pandas==1.0.5
scikit-learn==0.23.1
scipy==1.4.1
tqdm==4.47.0
nltk==3.5
emoji==0.6.0
transformers==3.0.2
ekphrasis==0.5.1
tweet-preprocessor==0.6.0
contractions==0.0.52
unidecode==1.2.0
requests>=2.11.1
flask==1.1.2
beautifulsoup4==4.9.3
python-dotenv==0.18.0
pyngrok==5.0.5
python-telegram-bot==13.7
pyrebase4==4.5.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.6.0+cpu
torchvision==0.7.0+cpu
